Simple.kt -> SimpleTypes.kt, make Null an object, relocate String to Collections.kt and rename it to TextString
